+++ packages/paste-website/src/components/site-wrapper/SiteHeader.tsx
@@ -1,18 +1,19 @@
 import * as React from 'react';
 import {trackCustomEvent} from 'gatsby-plugin-google-analytics';
-import {Box} from '@twilio-paste/box';
-import {Stack} from '@twilio-paste/stack';
 import {Anchor} from '@twilio-paste/anchor';
+import {Box} from '@twilio-paste/box';
+import {Flex} from '@twilio-paste/flex';
 import {FormInput} from '@twilio-paste/form';
+import {Stack} from '@twilio-paste/stack';
+import {Text} from '@twilio-paste/text';
 import {useTheme} from '@twilio-paste/theme';
+import {SearchIcon} from '@twilio-paste/icons/esm/SearchIcon';
 import {ThemeSwitcher} from '../ThemeSwitcher';
-import {ContactUsMenu} from '../ContactUsMenu';
 import GithubIcon from '../icons/GithubIcon';
-import {getMainContentComputedOffset} from '../../utils/stickyUtils';
-import {SiteBodyContext} from './SiteBodyContext';
+import {SiteHeaderLogo} from './SiteHeaderLogo';
+import {PSA_ALERT_HEIGHT} from '../../constants';
 
 export const SiteHeader: React.FC = () => {
-  const {isPasteTheme} = React.useContext(SiteBodyContext);
   const theme = useTheme();
   const [value, setValue] = React.useState('');
 
@@ -29,35 +30,46 @@ export const SiteHeader: React.FC = () => {
   return (
     <Box
       as="aside"
-      backgroundColor="colorBackgroundBody"
-      borderColor="colorBorderLight"
-      borderStyle="none none solid none"
-      borderBottomWidth="borderWidth10"
-      paddingLeft="space200"
-      paddingRight="space80"
-      paddingTop="space50"
-      paddingBottom="space50"
+      display="flex"
+      alignItems="center"
+      backgroundColor="colorBackgroundPrimaryDarkest"
       position="sticky"
-      top={`${getMainContentComputedOffset(isPasteTheme)}px`}
+      top={`${PSA_ALERT_HEIGHT}px`}
       zIndex="zIndex80"
     >
-      <Box display="flex" justifyContent="space-between" alignItems="center" flexDirection="row">
+      <SiteHeaderLogo siteTitle="Paste" siteSubTitle="Design System" />
+      <Box
+        display="flex"
+        justifyContent="space-between"
+        alignItems="center"
+        flexDirection="row"
+        borderLeftColor="colorBorderInverseDarker"
+        borderLeftStyle="solid"
+        borderLeftWidth="borderWidth10"
+        paddingTop="space60"
+        paddingRight="space70"
+        paddingBottom="space60"
+        paddingLeft="space70"
+        width="100%"
+      >
         <ThemeSwitcher />
         <Box marginTop="space0">
           <Stack orientation="horizontal" spacing="space60">
             <Box minWidth={['size30', 'size30', 'size30', 'size40']}>
               <FormInput
                 id="search-input"
-                type="search"
+                type="text"
                 placeholder="Search..."
                 value={value}
                 onChange={event => {
                   setValue(event.target.value);
                 }}
+                variant="inverse"
+                insertBefore={<SearchIcon decorative />}
               />
             </Box>
-            <ContactUsMenu />
             <Anchor
+              display="block"
               href="https://www.github.com/twilio-labs/paste"
               onClick={() =>
                 trackCustomEvent({
@@ -66,12 +78,18 @@ export const SiteHeader: React.FC = () => {
                   label: 'Github Icon',
                 })
               }
+              variant="inverse"
             >
-              <GithubIcon
-                css={{height: theme.heights.sizeIcon30, width: theme.heights.sizeIcon30}}
-                title="View this project on github"
-                color={theme.textColors.colorTextWeak}
-              />
+              <Flex as="span" vAlignContent="center">
+                <GithubIcon
+                  css={{height: theme.heights.sizeIcon30, width: theme.heights.sizeIcon30}}
+                  title="View this project on github"
+                  color={theme.textColors.colorTextIconInverse}
+                />
+                <Text as="span" color="inherit" marginLeft="space30">
+                  GitHub
+                </Text>
+              </Flex>
             </Anchor>
           </Stack>
         </Box>