+++ packages/paste-website/src/components/site-wrapper/sidebar/SidebarNavigation.tsx
@@ -1,14 +1,16 @@
 import * as React from 'react';
 import {graphql, useStaticQuery} from 'gatsby';
+import {Box} from '@twilio-paste/box';
 import {
-  SiteNav,
-  SiteNavList,
-  SiteNavNestList,
-  SiteNavItem,
-  SiteNavAnchor,
-  SiteNavButton,
-  SiteNavAnchorArrow,
-} from './SidebarNavigation.styles';
+  useDisclosurePrimitiveState,
+  DisclosurePrimitive,
+  DisclosurePrimitiveContent,
+} from '@twilio-paste/disclosure-primitive';
+import {SidebarAnchor} from './SidebarAnchor';
+import {SidebarDisclosureButton} from './SidebarDisclosureButton';
+import {SidebarItem} from './SidebarItem';
+import {SidebarNestedItem} from './SidebarNestedItem';
+import {SidebarNestedList} from './SidebarNestedList';
 import {PackageStatus, SidebarCategoryRoutes} from '../../../constants';
 import {getCurrentPathname, getNameFromPackageName, getHumanizedNameFromPackageName} from '../../../utils/RouteUtils';
 import {filteredComponents} from '../../../utils/componentFilters';
@@ -99,223 +101,254 @@ const pageQuery = graphql`
 const SidebarNavigation: React.FC<SidebarNavigationProps> = () => {
   const data: SiteWrapperPageQuery = useStaticQuery(pageQuery);
 
-  const [componentsOpen, setComponentsOpen] = React.useState(
-    getCurrentPathname().startsWith(SidebarCategoryRoutes.COMPONENTS)
-  );
+  const gettingStartedDisclosure = useDisclosurePrimitiveState({
+    visible: getCurrentPathname().startsWith(SidebarCategoryRoutes.GETTING_STARTED),
+  });
 
-  const [primitivesOpen, setPrimitivesOpen] = React.useState(
-    getCurrentPathname().startsWith(SidebarCategoryRoutes.PRIMITIVES)
-  );
+  const componentsDisclosure = useDisclosurePrimitiveState({
+    visible: getCurrentPathname().startsWith(SidebarCategoryRoutes.COMPONENTS),
+  });
 
-  const [layoutOpen, setLayoutOpen] = React.useState(getCurrentPathname().startsWith(SidebarCategoryRoutes.LAYOUT));
+  const primitivesDisclosure = useDisclosurePrimitiveState({
+    visible: getCurrentPathname().startsWith(SidebarCategoryRoutes.PRIMITIVES),
+  });
 
-  const [iconSystemOpen, setIconSystemOpen] = React.useState(
-    getCurrentPathname().startsWith(SidebarCategoryRoutes.ICON_SYSTEM)
-  );
+  const layoutDisclosure = useDisclosurePrimitiveState({
+    visible: getCurrentPathname().startsWith(SidebarCategoryRoutes.LAYOUT),
+  });
 
-  const [contentOpen, setContentOpen] = React.useState(getCurrentPathname().startsWith(SidebarCategoryRoutes.CONTENT));
+  const iconDisclosure = useDisclosurePrimitiveState({
+    visible: getCurrentPathname().startsWith(SidebarCategoryRoutes.ICON_SYSTEM),
+  });
 
-  const [gettingStartedOpen, setgettingStartedOpen] = React.useState(
-    getCurrentPathname().startsWith(SidebarCategoryRoutes.GETTING_STARTED)
-  );
+  const contentDisclosure = useDisclosurePrimitiveState({
+    visible: getCurrentPathname().startsWith(SidebarCategoryRoutes.CONTENT),
+  });
 
-  const [tokensOpen, setTokensOpen] = React.useState(getCurrentPathname().startsWith(SidebarCategoryRoutes.TOKENS));
+  const tokensDisclosure = useDisclosurePrimitiveState({
+    visible: getCurrentPathname().startsWith(SidebarCategoryRoutes.TOKENS),
+  });
 
   return (
-    <SiteNav role="navigation" aria-label="Main">
-      <SiteNavList>
-        <SiteNavItem>
-          <SiteNavAnchor to="/">Home</SiteNavAnchor>
-        </SiteNavItem>
-        <SiteNavItem>
-          <SiteNavButton
-            onClick={() => setgettingStartedOpen(!gettingStartedOpen)}
-            isOpen={gettingStartedOpen}
-            aria-expanded={gettingStartedOpen}
-          >
+    <Box as="nav" marginTop="space70" overflow="auto" role="navigation" aria-label="Main">
+      <Box as="ul" padding="space0" margin="space0" listStyleType="none">
+        <SidebarItem>
+          <SidebarAnchor to="/">Home</SidebarAnchor>
+        </SidebarItem>
+        <SidebarItem>
+          <DisclosurePrimitive as={SidebarDisclosureButton} {...gettingStartedDisclosure}>
             Getting Started
-            <SiteNavAnchorArrow isOpen={gettingStartedOpen} />
-          </SiteNavButton>
-          <SiteNavNestList isOpen={gettingStartedOpen}>
-            <SiteNavItem>
-              <SiteNavAnchor to="/getting-started/engineering">Engineering Guidelines</SiteNavAnchor>
-            </SiteNavItem>
-            <SiteNavItem>
-              <SiteNavAnchor to="/getting-started/design">Design Guidelines</SiteNavAnchor>
-            </SiteNavItem>
-            <SiteNavItem>
-              <SiteNavAnchor to="/getting-started/working-guide">How to work with us</SiteNavAnchor>
-            </SiteNavItem>
-            <SiteNavItem>
-              <SiteNavAnchor to="/getting-started/about-paste">About Paste</SiteNavAnchor>
-            </SiteNavItem>
-            <SiteNavItem>
-              <SiteNavAnchor to="/getting-started/faqs">FAQs</SiteNavAnchor>
-            </SiteNavItem>
-          </SiteNavNestList>
-        </SiteNavItem>
-        <SiteNavItem>
-          <SiteNavAnchor to="/principles">Principles</SiteNavAnchor>
-        </SiteNavItem>
-        <SiteNavItem>
-          <SiteNavButton onClick={() => setTokensOpen(!tokensOpen)} isOpen={tokensOpen} aria-expanded={tokensOpen}>
+          </DisclosurePrimitive>
+          <DisclosurePrimitiveContent {...gettingStartedDisclosure}>
+            <SidebarNestedList>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to="/getting-started/engineering">
+                  Engineering Guidelines
+                </SidebarAnchor>
+              </SidebarNestedItem>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to="/getting-started/design">
+                  Design Guidelines
+                </SidebarAnchor>
+              </SidebarNestedItem>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to="/getting-started/working-guide">
+                  How to work with us
+                </SidebarAnchor>
+              </SidebarNestedItem>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to="/getting-started/about-paste">
+                  About Paste
+                </SidebarAnchor>
+              </SidebarNestedItem>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to="/getting-started/faqs">
+                  FAQs
+                </SidebarAnchor>
+              </SidebarNestedItem>
+            </SidebarNestedList>
+          </DisclosurePrimitiveContent>
+        </SidebarItem>
+        <SidebarItem>
+          <SidebarAnchor to="/principles">Principles</SidebarAnchor>
+        </SidebarItem>
+        <SidebarItem>
+          <DisclosurePrimitive as={SidebarDisclosureButton} {...tokensDisclosure}>
             Design Tokens
-            <SiteNavAnchorArrow isOpen={tokensOpen} />
-          </SiteNavButton>
-          <SiteNavNestList isOpen={tokensOpen}>
-            <SiteNavItem>
-              <SiteNavAnchor to={SidebarCategoryRoutes.TOKENS}>Token list</SiteNavAnchor>
-            </SiteNavItem>
-            <SiteNavItem>
-              <SiteNavAnchor to={`${SidebarCategoryRoutes.TOKENS}/how-to-compose-custom-ui-with-tokens`}>
-                How to compose custom UI with tokens
-              </SiteNavAnchor>
-            </SiteNavItem>
-            <SiteNavItem>
-              <SiteNavAnchor to={`${SidebarCategoryRoutes.TOKENS}/design-tokens-package`}>
-                Design Tokens package
-              </SiteNavAnchor>
-            </SiteNavItem>
-            <SiteNavItem>
-              <SiteNavAnchor to={`${SidebarCategoryRoutes.TOKENS}/theme-package`}>Theme package</SiteNavAnchor>
-            </SiteNavItem>
-          </SiteNavNestList>
-        </SiteNavItem>
-        <SiteNavItem>
-          <SiteNavButton
-            onClick={() => setComponentsOpen(!componentsOpen)}
-            isOpen={componentsOpen}
-            aria-expanded={componentsOpen}
-          >
+          </DisclosurePrimitive>
+          <DisclosurePrimitiveContent {...tokensDisclosure}>
+            <SidebarNestedList>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to={SidebarCategoryRoutes.TOKENS}>
+                  Token list
+                </SidebarAnchor>
+              </SidebarNestedItem>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to={`${SidebarCategoryRoutes.TOKENS}/how-to-compose-custom-ui-with-tokens`}>
+                  How to compose custom UI with tokens
+                </SidebarAnchor>
+              </SidebarNestedItem>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to={`${SidebarCategoryRoutes.TOKENS}/design-tokens-package`}>
+                  Design Tokens package
+                </SidebarAnchor>
+              </SidebarNestedItem>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to={`${SidebarCategoryRoutes.TOKENS}/theme-package`}>
+                  Theme package
+                </SidebarAnchor>
+              </SidebarNestedItem>
+            </SidebarNestedList>
+          </DisclosurePrimitiveContent>
+        </SidebarItem>
+        <SidebarItem>
+          <DisclosurePrimitive as={SidebarDisclosureButton} {...componentsDisclosure}>
             Components
-            <SiteNavAnchorArrow isOpen={componentsOpen} />
-          </SiteNavButton>
-          <SiteNavNestList isOpen={componentsOpen}>
-            <SiteNavItem>
-              <SiteNavAnchor to={SidebarCategoryRoutes.COMPONENTS}>Overview</SiteNavAnchor>
-            </SiteNavItem>
-            {data.allPasteComponent.edges.filter(filteredComponents).map(({node}) => {
-              return (
-                <SiteNavItem key={node.name}>
-                  <SiteNavAnchor to={`${SidebarCategoryRoutes.COMPONENTS}/${getNameFromPackageName(node.name)}`}>
-                    {getHumanizedNameFromPackageName(node.name)}
-                  </SiteNavAnchor>
-                </SiteNavItem>
-              );
-            })}
-          </SiteNavNestList>
-        </SiteNavItem>
-        <SiteNavItem>
-          <SiteNavButton
-            onClick={() => setPrimitivesOpen(!primitivesOpen)}
-            isOpen={primitivesOpen}
-            aria-expanded={primitivesOpen}
-          >
+          </DisclosurePrimitive>
+          <DisclosurePrimitiveContent {...componentsDisclosure}>
+            <SidebarNestedList>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to={SidebarCategoryRoutes.COMPONENTS}>
+                  Overview
+                </SidebarAnchor>
+              </SidebarNestedItem>
+              {data.allPasteComponent.edges.filter(filteredComponents).map(({node}) => {
+                return (
+                  <SidebarNestedItem key={node.name}>
+                    <SidebarAnchor
+                      nested
+                      to={`${SidebarCategoryRoutes.COMPONENTS}/${getNameFromPackageName(node.name)}`}
+                    >
+                      {getHumanizedNameFromPackageName(node.name)}
+                    </SidebarAnchor>
+                  </SidebarNestedItem>
+                );
+              })}
+            </SidebarNestedList>
+          </DisclosurePrimitiveContent>
+        </SidebarItem>
+        <SidebarItem>
+          <DisclosurePrimitive as={SidebarDisclosureButton} {...primitivesDisclosure}>
             Primitives
-            <SiteNavAnchorArrow isOpen={primitivesOpen} />
-          </SiteNavButton>
-          <SiteNavNestList isOpen={primitivesOpen}>
-            <SiteNavItem>
-              <SiteNavAnchor to={SidebarCategoryRoutes.PRIMITIVES}>Overview</SiteNavAnchor>
-            </SiteNavItem>
-            {data.allPastePrimitive.edges
-              .filter(({node}) => node.status !== PackageStatus.BACKLOG)
-              .map(({node}) => {
-                return (
-                  <SiteNavItem key={node.name}>
-                    <SiteNavAnchor to={`${SidebarCategoryRoutes.PRIMITIVES}/${getNameFromPackageName(node.name)}`}>
-                      {getHumanizedNameFromPackageName(node.name)}
-                    </SiteNavAnchor>
-                  </SiteNavItem>
-                );
-              })}
-          </SiteNavNestList>
-        </SiteNavItem>
-        <SiteNavItem>
-          <SiteNavButton onClick={() => setLayoutOpen(!layoutOpen)} isOpen={layoutOpen} aria-expanded={layoutOpen}>
+          </DisclosurePrimitive>
+          <DisclosurePrimitiveContent {...primitivesDisclosure}>
+            <SidebarNestedList>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to={SidebarCategoryRoutes.PRIMITIVES}>
+                  Overview
+                </SidebarAnchor>
+              </SidebarNestedItem>
+              {data.allPastePrimitive.edges
+                .filter(({node}) => node.status !== PackageStatus.BACKLOG)
+                .map(({node}) => {
+                  return (
+                    <SidebarNestedItem key={node.name}>
+                      <SidebarAnchor
+                        nested
+                        to={`${SidebarCategoryRoutes.PRIMITIVES}/${getNameFromPackageName(node.name)}`}
+                      >
+                        {getHumanizedNameFromPackageName(node.name)}
+                      </SidebarAnchor>
+                    </SidebarNestedItem>
+                  );
+                })}
+            </SidebarNestedList>
+          </DisclosurePrimitiveContent>
+        </SidebarItem>
+        <SidebarItem>
+          <DisclosurePrimitive as={SidebarDisclosureButton} {...layoutDisclosure}>
             Layout
-            <SiteNavAnchorArrow isOpen={layoutOpen} />
-          </SiteNavButton>
-          <SiteNavNestList isOpen={layoutOpen}>
-            <SiteNavItem>
-              <SiteNavAnchor to={SidebarCategoryRoutes.LAYOUT}>Overview</SiteNavAnchor>
-            </SiteNavItem>
-            {data.allPasteLayout.edges
-              .filter(({node}) => node.status !== PackageStatus.BACKLOG)
-              .map(({node}) => {
-                return (
-                  <SiteNavItem key={node.name}>
-                    <SiteNavAnchor to={`${SidebarCategoryRoutes.LAYOUT}/${getNameFromPackageName(node.name)}`}>
-                      {getHumanizedNameFromPackageName(node.name)}
-                    </SiteNavAnchor>
-                  </SiteNavItem>
-                );
-              })}
-          </SiteNavNestList>
-        </SiteNavItem>
-        <SiteNavItem>
-          <SiteNavButton
-            onClick={() => setIconSystemOpen(!iconSystemOpen)}
-            isOpen={iconSystemOpen}
-            aria-expanded={iconSystemOpen}
-          >
+          </DisclosurePrimitive>
+          <DisclosurePrimitiveContent {...layoutDisclosure}>
+            <SidebarNestedList>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to={SidebarCategoryRoutes.LAYOUT}>
+                  Overview
+                </SidebarAnchor>
+              </SidebarNestedItem>
+              {data.allPasteLayout.edges
+                .filter(({node}) => node.status !== PackageStatus.BACKLOG)
+                .map(({node}) => {
+                  return (
+                    <SidebarNestedItem key={node.name}>
+                      <SidebarAnchor nested to={`${SidebarCategoryRoutes.LAYOUT}/${getNameFromPackageName(node.name)}`}>
+                        {getHumanizedNameFromPackageName(node.name)}
+                      </SidebarAnchor>
+                    </SidebarNestedItem>
+                  );
+                })}
+            </SidebarNestedList>
+          </DisclosurePrimitiveContent>
+        </SidebarItem>
+        <SidebarItem>
+          <DisclosurePrimitive as={SidebarDisclosureButton} {...iconDisclosure}>
             Icons
-            <SiteNavAnchorArrow isOpen={iconSystemOpen} />
-          </SiteNavButton>
-          <SiteNavNestList isOpen={iconSystemOpen}>
-            <SiteNavItem>
-              <SiteNavAnchor to={`${SidebarCategoryRoutes.ICON_SYSTEM}`}>Icons List</SiteNavAnchor>
-            </SiteNavItem>
-            <SiteNavItem>
-              <SiteNavAnchor to={`${SidebarCategoryRoutes.ICON_SYSTEM}/usage-guidelines`}>
-                Usage Guidelines
-              </SiteNavAnchor>
-            </SiteNavItem>
-            <SiteNavItem>
-              <SiteNavAnchor to={`${SidebarCategoryRoutes.ICON_SYSTEM}/how-to-add-an-icon`}>
-                How to add an icon
-              </SiteNavAnchor>
-            </SiteNavItem>
-          </SiteNavNestList>
-        </SiteNavItem>
-        <SiteNavItem>
-          <SiteNavAnchor to="/illustrations">Illustrations</SiteNavAnchor>
-        </SiteNavItem>
-        <SiteNavItem>
-          <SiteNavButton onClick={() => setContentOpen(!contentOpen)} isOpen={contentOpen} aria-expanded={contentOpen}>
+          </DisclosurePrimitive>
+          <DisclosurePrimitiveContent {...iconDisclosure}>
+            <SidebarNestedList>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to={`${SidebarCategoryRoutes.ICON_SYSTEM}`}>
+                  Icons List
+                </SidebarAnchor>
+              </SidebarNestedItem>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to={`${SidebarCategoryRoutes.ICON_SYSTEM}/usage-guidelines`}>
+                  Usage Guidelines
+                </SidebarAnchor>
+              </SidebarNestedItem>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to={`${SidebarCategoryRoutes.ICON_SYSTEM}/how-to-add-an-icon`}>
+                  How to add an icon
+                </SidebarAnchor>
+              </SidebarNestedItem>
+            </SidebarNestedList>
+          </DisclosurePrimitiveContent>
+        </SidebarItem>
+        <SidebarItem>
+          <SidebarAnchor to="/illustrations">Illustrations</SidebarAnchor>
+        </SidebarItem>
+        <SidebarItem>
+          <DisclosurePrimitive as={SidebarDisclosureButton} {...contentDisclosure}>
             Content
-            <SiteNavAnchorArrow isOpen={contentOpen} />
-          </SiteNavButton>
-          <SiteNavNestList isOpen={contentOpen}>
-            <SiteNavItem>
-              <SiteNavAnchor to={SidebarCategoryRoutes.CONTENT}>Guideline Usage</SiteNavAnchor>
-            </SiteNavItem>
-            <SiteNavItem>
-              <SiteNavAnchor to={`${SidebarCategoryRoutes.CONTENT}/voice-and-tone`}>Voice and tone</SiteNavAnchor>
-            </SiteNavItem>
-            <SiteNavItem>
-              <SiteNavAnchor to={`${SidebarCategoryRoutes.CONTENT}/product-style-guide`}>
-                Product style guide
-              </SiteNavAnchor>
-            </SiteNavItem>
-            <SiteNavItem>
-              <SiteNavAnchor to={`${SidebarCategoryRoutes.CONTENT}/writing-for-components`}>
-                Writing for components
-              </SiteNavAnchor>
-            </SiteNavItem>
-            <SiteNavItem>
-              <SiteNavAnchor to={`${SidebarCategoryRoutes.CONTENT}/get-support`}>Get support</SiteNavAnchor>
-            </SiteNavItem>
-          </SiteNavNestList>
-        </SiteNavItem>
-        <SiteNavItem>
-          <SiteNavAnchor to="/libraries">Libraries</SiteNavAnchor>
-        </SiteNavItem>
-        <SiteNavItem>
-          <SiteNavAnchor to="/roadmap">Roadmap</SiteNavAnchor>
-        </SiteNavItem>
-      </SiteNavList>
-    </SiteNav>
+          </DisclosurePrimitive>
+          <DisclosurePrimitiveContent {...contentDisclosure}>
+            <SidebarNestedList>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to={SidebarCategoryRoutes.CONTENT}>
+                  Guideline Usage
+                </SidebarAnchor>
+              </SidebarNestedItem>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to={`${SidebarCategoryRoutes.CONTENT}/voice-and-tone`}>
+                  Voice and tone
+                </SidebarAnchor>
+              </SidebarNestedItem>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to={`${SidebarCategoryRoutes.CONTENT}/product-style-guide`}>
+                  Product style guide
+                </SidebarAnchor>
+              </SidebarNestedItem>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to={`${SidebarCategoryRoutes.CONTENT}/writing-for-components`}>
+                  Writing for components
+                </SidebarAnchor>
+              </SidebarNestedItem>
+              <SidebarNestedItem>
+                <SidebarAnchor nested to={`${SidebarCategoryRoutes.CONTENT}/get-support`}>
+                  Get support
+                </SidebarAnchor>
+              </SidebarNestedItem>
+            </SidebarNestedList>
+          </DisclosurePrimitiveContent>
+        </SidebarItem>
+        <SidebarItem>
+          <SidebarAnchor to="/libraries">Libraries</SidebarAnchor>
+        </SidebarItem>
+        <SidebarItem>
+          <SidebarAnchor to="/roadmap">Roadmap</SidebarAnchor>
+        </SidebarItem>
+      </Box>
+    </Box>
   );
 };