+++ packages/paste-core/components/menu/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.31](https://github.com/twilio-labs/paste/compare/@twilio-paste/menu@1.1.30...@twilio-paste/menu@1.1.31) (2020-10-07)
+
+**Note:** Version bump only for package @twilio-paste/menu
+
+
+
+
+
 ## [1.1.30](https://github.com/twilio-labs/paste/compare/@twilio-paste/menu@1.1.29...@twilio-paste/menu@1.1.30) (2020-10-07)
 
 **Note:** Version bump only for package @twilio-paste/menu