+++ packages/paste-core/components/modal/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.30](https://github.com/twilio-labs/paste/compare/@twilio-paste/modal@1.0.29...@twilio-paste/modal@1.0.30) (2020-10-07)
+
+**Note:** Version bump only for package @twilio-paste/modal
+
+
+
+
+
 ## [1.0.29](https://github.com/twilio-labs/paste/compare/@twilio-paste/modal@1.0.28...@twilio-paste/modal@1.0.29) (2020-10-07)
 
 **Note:** Version bump only for package @twilio-paste/modal