+++ packages/paste-core/components/paragraph/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.46](https://github.com/twilio-labs/paste/compare/@twilio-paste/paragraph@1.1.45...@twilio-paste/paragraph@1.1.46) (2020-10-07)
+
+**Note:** Version bump only for package @twilio-paste/paragraph
+
+
+
+
+
 ## [1.1.45](https://github.com/twilio-labs/paste/compare/@twilio-paste/paragraph@1.1.44...@twilio-paste/paragraph@1.1.45) (2020-10-07)
 
 **Note:** Version bump only for package @twilio-paste/paragraph