+++ packages/paste-core/primitives/sibling-box/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.3](https://github.com/twilio-labs/paste/compare/@twilio-paste/sibling-box@1.1.2...@twilio-paste/sibling-box@1.1.3) (2020-10-07)
+
+**Note:** Version bump only for package @twilio-paste/sibling-box
+
+
+
+
+
 ## [1.1.2](https://github.com/twilio-labs/paste/compare/@twilio-paste/sibling-box@1.1.1...@twilio-paste/sibling-box@1.1.2) (2020-10-07)
 
 **Note:** Version bump only for package @twilio-paste/sibling-box